requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
tenacity==8.5.0
fastapi-cache2==0.2.2
redis==5.0.8
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.43
alembic==1.13.2
asyncpg==0.29.0
psycopg[binary]==3.2.1
greenlet==3.0.3
requests==2.32.3
jinja2
python-multipart
apscheduler
python-jose[cryptography]
email-validator
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
praw>=7.7,<8
aiohttp>=3.9.0

